requirements.txt
appdirs==1.4.3
asgiref==3.3.1
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.1.6
django-admin-rangefilter==0.6.4
django-baton==2.1.3
django-dynamic-raw-id==2.8
django-extensions==3.1.1
django-jsignature==0.10
django-logentry-admin==1.0.6
django-mapbox-location-field==1.7.1
django-model-utils==4.1.1
djangorestframework==3.12.2
flake8==3.8.4
google-api-core==1.26.3
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
html5lib==1.0.1
httplib2==0.19.1
idna==2.10
ipaddr==2.2.0
isort==5.7.0
lockfile==0.12.2
mccabe==0.6.1
msgpack==0.6.2
#numpy==1.20.2
packaging==20.9
pep517==0.8.2
Pillow==8.1.2
pip-chill==1.0.1
progress==1.5
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytoml==0.1.21
pytz==2021.1
requests==2.25.1
retrying==1.3.3
reverse-geocoder==1.5.1
rsa==4.7.2
scipy==1.6.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
webencodings==0.5.1
yapf==0.30.0
